+++ app/src/main/java/ru/sibur/android/garbagecollector/Storage.java
@@ -5,9 +5,9 @@ package ru.sibur.android.garbagecollector;
  */
 
 public abstract class Storage {
-    abstract void addMoney(int amount);
-    abstract int getMoney();
-    abstract int getTotalMoneyEarned();
+    abstract void addMoney(long amount);
+    abstract long getMoney();
+    abstract long getTotalMoneyEarned();
     abstract void incrementShopItemCount (String itemCountKey);
     abstract int getShopItemCount (String itemCountKey);
